requirements.txt
mysql-connector-python
numpy
pandas
matplotlib
scipy
joblib
scikit-learn
imbalanced-learn
configparser
sshtunnel
openpyxl